% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_tools.R
\name{as_cell}
\alias{as_cell}
\title{Convert points to 1 point per cell}
\usage{
as_cell(pts, base_rast = NULL, res = 0.05)
}
\arguments{
\item{pts}{dataframe with two columns named decimalLongitude and decimalLatitude}

\item{base_rast}{an optional raster (SpatRaster format) to be used as a model.
If supplied, the results are masked according to this raster, and any point that falls
outside a valid area will be removed.}

\item{res}{if \code{base_rast} is not supplied, then a raster with the given resolution
(in degrees) is used as a base}
}
\value{
data.frame with points
}
\description{
Convert points to 1 point per cell
}
\examples{
\dontrun{
as_cell(all_pts)
}
}
