% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_tools.R
\name{cogeo_optim}
\alias{cogeo_optim}
\title{Convert GeoTIFF files to Cloud Optimized GeoTIFF format (COG)}
\usage{
cogeo_optim(id, remove_files = TRUE, verbose = FALSE)
}
\arguments{
\item{id}{path to the GeoTIFF file to be optimized}

\item{remove_files}{if \code{TRUE} (default), the original file is removed if the
optimization succeeds or the COG is removed if the optimization failed
somehow}

\item{verbose}{if \code{TRUE} output messages}
}
\value{
optimization status and saved files
}
\description{
Convert GeoTIFF files to Cloud Optimized GeoTIFF format (COG)
}
\details{
There are three possible status returned by the function:
\itemize{
\item optimization-succeeded: optimization was successful and file is saved.
The non-optimized file is removed if \code{remove_files = TRUE}.
\item optimization-failed: optimization failed.
\item optimization-gen-failed: file was probably generated, but by some reason
it was not correctly optimized. This was included more for control, as such a
case should be extremely rare.
}

This function depends on the \href{https://cogeotiff.github.io/rio-cogeo/}{rio-cogeo plugin}, which
is a Python program with CLI interface. To install the plugin, you should use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{$ pip install rio-cogeo
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
cogeo_optim("test_file.tif")
}
}
