% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{cv_mod}
\alias{cv_mod}
\title{Cross-validate model}
\usage{
cv_mod(
  n_folds,
  folds,
  labels,
  training,
  weights = NULL,
  prob_mode = F,
  metrics,
  normalize = FALSE,
  pred_type,
  model_fun,
  return_block_pred = FALSE,
  ...
)
}
\arguments{
\item{n_folds}{the number of folds}

\item{folds}{folds index}

\item{labels}{the presence-absence vector}

\item{training}{training data}

\item{metrics}{which metrics return}

\item{normalize}{rescale the predicted values to the scale 0-1. Relevant for
down-weighted or infinitely weighted models.}

\item{pred_type}{type of prediction}

\item{model_fun}{the function of the model}

\item{return_block_pred}{if \code{TRUE}, then the predictions for each block
are returned.}

\item{...}{additional parameters for the model function}
}
\value{
a list with the metrics for each fold
}
\description{
Cross-validate model
}
\examples{
\dontrun{

}
}
