% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_get_occ_local.R
\name{.mp_get_occ_local}
\alias{.mp_get_occ_local}
\title{Get occurrence data from a local dataset}
\usage{
.mp_get_occ_local(
  loc_file,
  scientificName,
  startdate = NULL,
  enddate = NULL,
  geometry = NULL
)
}
\arguments{
\item{loc_file}{the absolute or relative path to the local dataset file}

\item{scientificName}{the name of the species}

\item{startdate}{an optional start date to filter occurrences (YYYY-MM-DD)}

\item{enddate}{an optional end date to filter occurrences (YYYY-MM-DD)}

\item{geometry}{an optional geometry (WKT format) to filter occurrences}
}
\value{
the filterd locations
}
\description{
Get occurrence data from local datasets. Those have to contain, at least, 
one collumn called "scientificName" and two columns with longitude/latitude 
data coded as lon/lat, x/y or decimalLongitude/decimalLatitude. 
Date columns should be named date. Additional columns are ignored.
}
