% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{ensemble_eval}
\alias{ensemble_eval}
\title{Evaluate ensemble}
\usage{
ensemble_eval(ensemble, sdm_data, ...)
}
\arguments{
\item{ensemble}{the SpatRaster object containing the ensemble for the current
period. If multiple layers are available, only the first one will be used.}

\item{sdm_data}{\code{sdm_dat} object containing coordinates of the occurrence
records and optionally evaluation dataset}

\item{...}{further arguments passed to \code{\link[=eval_metrics]{eval_metrics()}}}
}
\value{
data.frame
}
\description{
Evaluate ensemble
}
\examples{
\dontrun{
ensemble_eval(ensemble_layer, sp_data)
}
}
