% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{ensemble_models}
\alias{ensemble_models}
\title{Ensemble model predictions}
\usage{
ensemble_models(ens_method, ...)
}
\arguments{
\item{ens_method}{which method to use for ensemble (see details)}

\item{...}{two or more SpatRaster objects, or a list of two or more SpatRaster objects}
}
\value{
ensemble in SpatRaster format, with the metric and error layers
}
\description{
Ensemble model predictions
}
\details{
Three methods are currently implemented:
\itemize{
\item mean (average of model predictions, non-weighted)
\item median (median of model predictions)
\item committee averaging of binary predictions
}

Remember that for the last case, binary maps should be provided.
}
\examples{
\dontrun{
pred_ensemble <- ensemble_models("median", pred1, pred2)
}
}
