% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{ensemble_respcurves}
\alias{ensemble_respcurves}
\title{Ensemble response curves}
\usage{
ensemble_respcurves(ens_method, normalize, ...)
}
\arguments{
\item{ens_method}{which method to use for ensemble (see details)}

\item{normalize}{logical, if \code{TRUE},
then responses are normalized in the 0-1 scale before applying the ensemble}

\item{...}{two or more response curves objects generated with \code{\link[=resp_curves]{resp_curves()}}}
}
\value{
data.frame
}
\description{
Ensemble response curves
}
\details{
Three methods are currently implemented:
\itemize{
\item mean (average of model predictions, non-weighted)
\item median (median of model predictions)
\item committee averaging of binary predictions
}

Remember that for the last case, binary maps should be provided.
}
\examples{
\dontrun{
ensresp <- ensemble_respcurves("median", resp1, resp2)
}
}
