% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_tools.R
\name{find_file}
\alias{find_file}
\alias{parquet_to}
\alias{deband}
\title{Utilities for post-processing results}
\usage{
find_file(
  taxon,
  what = NULL,
  type = NULL,
  results_path = "results",
  model = NULL
)

parquet_to(file, to = "csv", new_path = NULL)

deband(raster_obj, sel_band, path = NULL, ...)
}
\arguments{
\item{taxon}{taxon ID (AphiaID) for the target species}

\item{what}{what you need (for example, "cvmetrics" will return files of
cross-validation metrics). If \code{NULL} or \code{NA} it will return all available
files}

\item{type}{optionally, provide the file type (example, "parquet")}

\item{results_path}{path to the results folder}

\item{model}{an optional character name for the model acronym. If not
provided it will use the most recent}

\item{file}{path to a parquet file}

\item{to}{target format. Can be one of "csv", "txt" or "rds"}

\item{new_path}{an optional path for the new file. If NULL, it will be saved
in the same path. Note that you should provide only the folder path, as the
file name will be exactly the same as the original file, but with the new
format}

\item{raster_obj}{a SpatRaster object or a path for a raster file (any format
accepted by \code{\link[terra:rast]{terra::rast()}})}

\item{sel_band}{a \code{character} with the band name or a \code{numeric} with the
number of the band}

\item{path}{an optional file path to save the individual band. This should be
a full path, including the file format (any accepted by
\code{\link[terra:writeRaster]{terra::writeRaster()}}). If \code{NULL}, returns the SpatRaster}

\item{...}{further arguments to \code{\link[terra:writeRaster]{terra::writeRaster()}}}
}
\value{
\code{find_file()} returns a character vector with file paths (full
address relative to the working directory)
\code{parquet_to()} returns nothing, save the file
\code{deband()} returns the SpatRaster if \code{path=NULL}, or nothing and saves the file
}
\description{
\code{find_file()} helps to find a specific file in the results folder for a
specific taxon

\code{parquet_to()} is a wrapper to quickly convert \code{parquet} files to other
common formats

\code{deband()} provides an easy way to extract a single band from multiband
rasters
}
\examples{
\dontrun{
files <- find_file(127326, "cvmetrics", type = "parquet")

parquet_to(files[1])

mask_file <- find_file(127326, "mask")

single_band <- deband(mask_file, sel_band = "fit_region")
}
}
