% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_prepare_blocks.R
\name{folds_info}
\alias{folds_info}
\title{Get information on the number of points per fold}
\usage{
folds_info(sdm_data, block_type = "spatial_grid")
}
\arguments{
\item{sdm_data}{an sdm_dat object prepared using \code{\link[=mp_prepare_data]{mp_prepare_data()}} and with
assigned blocks using \code{\link[=mp_prepare_blocks]{mp_prepare_blocks()}}}

\item{block_type}{the block type}
}
\value{
a data.frame containing number of points per class and per fold
}
\description{
Get information on the number of points per fold
}
\examples{
\dontrun{
folds_info(sdm_data)
}
}
