% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_get_conf.R
\name{gen_configuration}
\alias{gen_configuration}
\title{Generate configuration file in YAML format}
\usage{
gen_configuration(filename = "sdm_conf.yml", outdir = ".")
}
\arguments{
\item{filename}{a file name}

\item{outdir}{the output directory. If not existent, will be created}
}
\value{
saved file
}
\description{
This function will generate a configuration file in YAML format, which provides a
convenient way to save and retrieve configurations for running the SDMs.
}
\details{
Configurations from this file can be retrieved using the function \code{\link[=get_conf]{get_conf()}}.
For details on how to fill the file, see comments on the generated file.
}
\examples{
\dontrun{
gen_configuration()
}
}
