% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_tools.R
\name{gen_log}
\alias{gen_log}
\title{Generate log of modeling}
\usage{
gen_log(algos)
}
\arguments{
\item{algos}{algorithms that will be used to fit}
}
\value{
a list
}
\description{
Generate log of modeling
}
\examples{
sdm_log <- gen_log(c("maxent", "brt"))
\dontrun{
save_log(sdm_log, "species_log.json")
}
}
