% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_vsp.R
\name{gen_vsp}
\alias{gen_vsp}
\title{Generate virtual species for SDM testing}
\usage{
gen_vsp(
  layers,
  sim_funs = NULL,
  sim_pca = NULL,
  samp_n_low = 30,
  samp_n_high = 150,
  samp_rep = 10,
  samp_bias_layer = NULL,
  samp_bias_only = FALSE,
  samp_constr_shape = NULL,
  samp_pa = TRUE,
  samp_replace = FALSE,
  vsp_name = NULL,
  vsp_class = NULL,
  pred_fut = NULL,
  save_path = NULL,
  save_key = NULL,
  plot = TRUE,
  test_mode = FALSE
)
}
\arguments{
\item{layers}{environmental layers in raster* format (not terra compatible
yet).}

\item{sim_funs}{functional relationships for the virtual species, generated
with \code{\link[virtualspecies:formatFunctions]{virtualspecies::formatFunctions()}}. All layers should have a
corresponding function. If not supplied, then a PCA mode is used (then
sim_pca should be supplied).}

\item{sim_pca}{the type of niche simulated by the PCA model. Can be "narrow"
or "wide" (see details on \code{\link[virtualspecies:generateSpFromPCA]{virtualspecies::generateSpFromPCA()}}). Ignored
if \code{sim_funs} is supplied.}

\item{samp_n_low}{number of points for the low sampling scenario.}

\item{samp_n_high}{number of points for the highly sampled scenario.}

\item{samp_rep}{number of repetitions of the sampling.}

\item{samp_bias_layer}{an optional sample bias layer. Should be a raster with
0-1 values.}

\item{samp_bias_only}{in case that a sampling bias layer is supplied, define
if the sampling should be only with the bias or also without it. Default is
FALSE, meaning that sampling will occur with and without bias.}

\item{samp_constr_shape}{an optional shapefile to constrain the sampling to
that area.}

\item{samp_pa}{should absences also be returned? Interesting for comparisons
with presence-absence methods. When this option is set to \code{TRUE},
additional files with presence-absence are saved. Absences are sampled in
the same number as presences (equal prevalence).}

\item{samp_replace}{if \code{TRUE}, then the sampling of species occurs with
replacement and multiple points can be sampled on the same cell.}

\item{vsp_name}{a name for the virtual species. If \code{NULL} "Virtual
species" will be used.}

\item{vsp_class}{a 'class' for the virtual species. Helpful to save
information about the generating process. Can be for example
"vsp_pca_narrow". If \code{NULL}, "standard_vsp" is used.}

\item{pred_fut}{an optional named list of raster* with future environmental
conditions to predict the distribution of the species in future climates.
The names of the list will be used to save the future suitability rasters.}

\item{save_path}{where to save the files. If \code{NULL} the standard format
"data/virtual_species" will be used. Files will be saved according to the
key provided (see details).}

\item{save_key}{the key to save the species}

\item{plot}{plot distributions and graphics as the virtual species is
generated.}

\item{test_mode}{if set to \code{TRUE}, the function will stop before the
sampling of points. This is mainly intended for testing if the
environmental relationship functions are returning the expected suitability
pattern (in this case, plot = \code{TRUE}).}
}
\value{
saved files in the project format
}
\description{
Thin function is a wrapper around the \link{virtualspecies} package functions to
generate virtual species. The purpose of generating virtual species is to
test the capacity of distinct SDM algorithms to correctly return the "true"
distribution of species.
}
\details{
This function is a helper to generate virtual species for testing
SDM methods. Files will be saved in the path
data/virtual_species/key='save_key'/'file name' (note that the standard
folder is virtual_species, not species). A few points should be observed:
\itemize{
\item the \code{save_key} should be distinct from the AphiaIDs of the project species. You can use any number, but it's important to be consistent to avoid any confusion.
\item different than the others, the virtual species folder will contain more than only the species occurrence table.
\item the species occurrence information will be saved in the same way as the other species, but will contain only 6 collumns: key (save_key), scientificName (which will receive the vsp_name),
decimalLongitude, decimalLatitude, presence (1-0, denoting presence or
absence) and class, which will hold the vsp_class.
}

The files that will be saved in the folder are the following:
\itemize{
\item parquet files containing the virtual species occurrence for each sampling schema (very low, low and high)
\item if samp_pa is \code{TRUE}, the same files containing the absence info (saved as 'file*_pa.parquet')
\item raster files containing the suitability, presence-absence, probabilities of occurrence
\item if pred_fut is supplied, rasters containing the future distribution
}

NOTE: the process is random and you may want to set a seed before running the
function.
}
\examples{
\dontrun{
gen_vsp()
}
}
