% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_prepare_blocks.R
\name{get_block_grid}
\alias{get_block_grid}
\title{Get grid for spatial block CV}
\usage{
get_block_grid(
  sdm_data,
  env_layers,
  sel_vars = NULL,
  min_block_size = 0.2,
  max_block_size = 20,
  verbose = FALSE
)
}
\arguments{
\item{sdm_data}{object of type \code{sdm_dat} with species occurrence records}

\item{env_layers}{environmental layers in SpatRaster format}

\item{sel_vars}{which variables should be used for calculating the block size}

\item{min_block_size}{minimum size of the block. If the calculated range is
smaller than this threshold, then the minimum size is used instead}

\item{max_block_size}{maximum size of the block. If the calculated range is
larger than this threshold, then the maximum size is used instead}

\item{verbose}{print messages}
}
\value{
grid (SpatRaster)
}
\description{
This function uses the function \code{\link[blockCV:cv_spatial_autocor]{blockCV::cv_spatial_autocor()}} from the package
\link{blockCV} to calculate the appropriate size of the grid
}
\examples{
\dontrun{
get_block_grid(sp_data, env_data)
}
}
