% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_get_conf.R
\name{get_conf}
\alias{get_conf}
\title{Retrieve configuration settings from a configuration file}
\usage{
get_conf(conf_file = "sdm_conf.yml", what = c("groups", "variables"))
}
\arguments{
\item{conf_file}{the path for the configuration file}

\item{what}{what to retrieve (character vector)

This enable to retrieve configuration settings from a configuration YAML file
generated with \code{\link[=gen_configuration]{gen_configuration()}}. Although this function can be used, we recommend
using directly the following two functions:
\itemize{
\item \code{\link[=get_listbygroup]{get_listbygroup()}}: from the filters established in the "groups" section of
the configuration file, convert a list of species into a list with groups.
\item \code{\link[=get_envofgroup]{get_envofgroup()}}: for a certain group, read and load the environmental files
saved in a folder, following the list of chosen variables.
}}
}
\value{
a list of configurations
}
\description{
Retrieve configuration settings from a configuration file
}
\examples{
\dontrun{
configurations <- get_conf()
}
}
