% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_get_conf.R
\name{get_listbygroup}
\alias{get_listbygroup}
\title{Includes a new column of SDM groups to a species list}
\usage{
get_listbygroup(
  species_list,
  conf_file = "sdm_conf.yml",
  not_found_value = "NOT_FOUND"
)
}
\arguments{
\item{species_list}{the \code{data.frame} with the species list. Should have at least the
columns that were included as filters on the configuration file}

\item{conf_file}{the path for the configuration file}

\item{not_found_value}{which value use when the function is unable to establish a group}
}
\value{
the species list with a new column called "sdm_group"
}
\description{
This function will include a new grouping column to a species list (\code{data.frame} like objects)
based on the filters established in the configuration file.
}
\examples{
\dontrun{
new_list <- get_listbygroup(sp_list)
}
}
