% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_matching.R
\name{key_to_aphia}
\alias{key_to_aphia}
\title{Retrieve AphiaID from GBIF taxonKey}
\usage{
key_to_aphia(taxonkey, ...)
}
\arguments{
\item{taxonkey}{a vector of one or more GBIF taxonKeys}

\item{...}{additional parameters passed to \code{\link[=name_to_aphia]{name_to_aphia()}}}
}
\value{
dataframe with names
}
\description{
Search WoRMS AphiaID for a GBIF taxonKey. The function first gets the scientific
name for the provided key(s) and then search for the AphiaID(s) using
\code{\link[=name_to_aphia]{name_to_aphia()}}.
}
\examples{
\dontrun{
  key_to_aphia(2347567)
}
}
