% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_prepare_blocks.R
\name{manual_block}
\alias{manual_block}
\title{Fast spatial cross-validation blocks generation}
\usage{
manual_block(data_pts, user_grid = NULL, k, iterations = 50, legacy = FALSE)
}
\arguments{
\item{data_pts}{points in SF format. Should have a collumn named "presence"
containing the type of point (0 for background/absence and 1 for presence)}

\item{user_grid}{an user supplied SpatRaster (\link{terra}). If not supplied,
the function will generate a grid of resolution 1 degree based on the extent
of the points.}

\item{k}{number of folds}

\item{iterations}{number of iterations for sampling the points.}

\item{legacy}{if \code{TRUE} uses the old version of the function (see note)}
}
\value{
a list containing the folds, number of folds and division of records per fold.
}
\description{
This function will generate cross-validation blocks randomly based on a grid,
which may also be of latitudinal blocks. This function is simpler than the
\code{\link[blockCV:cv_spatial]{blockCV::cv_spatial()}}, but is aimed to be much faster.
}
\note{
In a previous version the function would sample 1:k blocks based on the number of
cells of the user_grid. While this works well, it adds a computational burden, because
just a part of those cells are actually necessary. In the new version it samples 1:k
blocks based on the cells at which a point is available. Since it is still completely random
no difference in the results is expected, except a speed gain. It also helps to
increase the chance of getting a good block division with less iterations.
}
\examples{
\dontrun{
user_grid <- rast(ext(-180, 180, -90, 90), resolution = 1)
blocks <- manual_block(species_points, user_grid, k = 5)
} 

}
