% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_bootstrap.R
\name{model_bootstrap}
\alias{model_bootstrap}
\title{Re-fit model with best parameter for bootstrap}
\usage{
model_bootstrap(sdm_data, algo, params, verbose = TRUE)
}
\arguments{
\item{sdm_data}{the \code{sdm_dat} object for model fitting}

\item{algo}{which algorithm to use for model fitting (see \code{\link[=sdm_options]{sdm_options()}})}

\item{params}{the best parameters. Should be a named list, with each name
equivalent to the parameters names from the \code{\link[=sdm_options]{sdm_options()}}. Only
one value for parameter is used, so passing multiple values for the same parameter
is ignored.}

\item{verbose}{if \code{TRUE}, print messages}
}
\value{
fitted model of class \code{sdm_result} or \code{sdm_esm_result} (which is a
list of \code{sdm_result} objects)
}
\description{
This function enable to re-fit any model using the best parameters and a new
dataset, intended mainly for bootstraping applications.
}
\examples{
\dontrun{
bp <- log_object$model_bestparams$maxent
# For maxent, it is a list within a list, so one more level
bp <- bp[[1]]
model_bootstrap(sdm_data, "maxent", params = bp)
}
}
