% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_get.R
\name{mp_get_local}
\alias{mp_get_local}
\title{Retrieve occurrence records from local file for individual species and save in the standard format}
\usage{
mp_get_local(
  local_file,
  database,
  sci_name = NULL,
  aphia_id = NULL,
  gbif_key = NULL,
  save_format = "parquet"
)
}
\arguments{
\item{local_file}{the relative or absolute path to the file (or folder in case of parquet datasets)}

\item{database}{the name of the database source, will be used to save the files}

\item{sci_name}{the scientific name of the species. If \code{NULL}, then the aphia_id or gbif_key will be used instead (recommended)}

\item{aphia_id}{the AphiaID of the species}

\item{gbif_key}{the GBIF taxon key}

\item{save_format}{the format to save the file (either "parquet" or "csv"; recommended is "parquet")}
}
\value{
individual file saved
}
\description{
Retrieve occurrence records from local file for individual species and save in the standard format
}
\examples{
\dontrun{

}
}
