% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_prepare_data.R
\name{mp_prepare_data}
\alias{mp_prepare_data}
\title{Prepare data for use in SDM}
\usage{
mp_prepare_data(
  training,
  eval_data = NULL,
  species_id,
  native_shp = NULL,
  env_layers,
  gen_quad = TRUE,
  quad_number = NULL,
  pred_quad = NULL,
  verbose = TRUE
)
}
\arguments{
\item{training}{a data frame containing xy coordinates  (and optionally a
column named presence with the presence-absence status) for the training
points (coordinates should be named decimalLongitude and decimalLatitude)}

\item{eval_data}{an optional data frame containing xy coordinates (and
optionally a column named presence with the presence-absence status) of the
points used for evaluation}

\item{species_id}{a name or id for the species data}

\item{native_shp}{an optional shapefile with the native or expert range data}

\item{env_layers}{environmental layers in terra rast format}

\item{gen_quad}{if \code{TRUE}, quadrature points are generated}

\item{quad_number}{the number of quadrature points to be generated. If set to
\code{NULL}, then 10000 points are generated}

\item{pred_quad}{optionally, you can supply a data.frame of quadrature points
containing the xy coordinates (named "decimalLongitude" and "decimalLatitude"),
a column named presence with values of 0, and the values for the variables.
Those should be named exactly as the \code{env_layers} and in the same order.
Supplying the pre-defined quadrature points can speed-up the process when several
runs will take place. If this is supplied, \code{gen_quad} is ignored}

\item{verbose}{should messages be printed}
}
\value{
an object of class sdm_dat containing the data for SDMs
}
\description{
Prepare data for use in SDM
}
\examples{
\dontrun{
  mp_prepare_data(training_data, env_layers = env)
}
}
