% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_download.R
\name{occurrence_gbif}
\alias{occurrence_gbif}
\title{Download marine data from GBIF}
\usage{
occurrence_gbif(
  scientificname = NULL,
  taxonid = NULL,
  startdate = NULL,
  enddate = NULL,
  startdepth = NULL,
  enddepth = NULL,
  geometry = NULL,
  absence = NULL,
  username = NULL,
  password = NULL,
  exclude = NULL,
  format = NULL,
  wait = TRUE,
  raw_path = NULL,
  import = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{scientificname}{the scientific name of the species or the taxon key for
the species obtained from GBIF. The second is preferred and you can obtain
the taxon key using the function \code{\link[=get_gbif_keys]{get_gbif_keys()}}. If a scientific name is
supplied, the function will first look in the GBIF taxonomic backbone for a
match and then use the key to download, but you will have no control on the
matching.}

\item{taxonid}{the AphiaID of the species. If supplied, then the function
will first look into WoRMS for the species name and then obtain the taxon
key from GBIF. Ignored if \code{scientificname} is supplied.}

\item{startdate}{the earliest date on which occurrence took place.}

\item{enddate}{the latest date on which the occurrence took place.}

\item{startdepth}{the minimum depth below the sea surface.}

\item{enddepth}{the maximum depth below the sea surface.}

\item{geometry}{a WKT geometry string (should not be too complex, or the
download will fail - see \code{\link[rgbif:occ_download]{rgbif::occ_download()}}).}

\item{absence}{only include absence records (\code{TRUE}), exclude absence
records (\code{NULL}) or include absence records (\code{include}).}

\item{username}{the GBIF username, if not stored on the R environment (see
details).}

\item{password}{the GBIF password, if not stored on the R environment (see
details).}

\item{exclude}{if \code{NULL} include all types of specimens. Alternatively,
you can pass a vector of types of specimen to exclude (e.g. supplying
\code{c("FOSSIL_SPECIMEN","LIVING_SPECIMEN")} will exclude fossils and
living specimens).}

\item{format}{format to download. One of \code{"SIMPLE_CSV"} or
\code{"SIMPLE_PARQUET"}. If \code{NULL} (default), then a simple CSV is
downloaded.}

\item{wait}{whether it should wait for the download to be completed. The
download from GBIF occurs in a different way then from OBIS. First, a call
is sent to the server, which will prepare the file for download. In
general, this takes just a few minutes, but depending on the size it can
take several minutes. If \code{wait} is \code{FALSE}, then the function
will return the download call, which can later be used with
[rgbif::occ_download_get())]. The default is \code{TRUE}, and the function
will wait until the download is ready and download it on the supplied
folder.

[rgbif::occ_download_get())]: R:rgbif::occ_download_get())}

\item{raw_path}{the path to the download folder. If \code{NULL}, the working
folder is used. Ignored if \code{wait} is \code{FALSE}.}

\item{import}{if \code{TRUE}, the data is imported as a data frame.}

\item{verbose}{if \code{TRUE} messages are printed.}
}
\value{
a data frame containing occurrences or a GBIF download object.
}
\description{
Download marine data from GBIF
}
\examples{
\dontrun{
occurrence_gbif("Leptuca thayeri")
}
}
