% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqc_get_info.R
\name{outqc_dist_tozarr}
\alias{outqc_dist_tozarr}
\title{Convert QC distances to Zarr format}
\usage{
outqc_dist_tozarr(distfolder, remove_distances = TRUE)
}
\arguments{
\item{distfolder}{the folder where the distances, calculated using \code{\link[=outqc_get_distances]{outqc_get_distances()}}, were saved}

\item{remove_distances}{if \code{TRUE}, after converting to Zarr the calculated distances are deleted,
and only the Zarr folder is kept}
}
\value{
saved Zarr files
}
\description{
Convert QC distances to Zarr format
}
\examples{
\dontrun{
outqc_dist_tozarr("distances")
}
}
