% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_dup_check.R
\name{outqc_dup_check}
\alias{outqc_dup_check}
\title{Remove duplicate datasets}
\usage{
outqc_dup_check(data_a, data_b = NULL, gh_precision = 6)
}
\arguments{
\item{data_a}{A named \code{list} of datasets which are to be compared or a
\code{data.frame}. Should have at least the collumns ScientificName,
decimalLongitude, decimalLatitude and year.}

\item{data_b}{(data.frame) If \code{data_a} is a \code{data.frame}, then \code{data_b}
should be supplied in order to produce a comparison between data providers.
Otherwise, the comparison will be only within the dataset.}

\item{gh_precision}{precision parameter for the geohash. (see details at
\code{\link[geohashTools:gh_encode]{geohashTools::gh_encode()}}).}
}
\value{
A \code{data.frame} with cleaned occurrence records.
}
\description{
Removes duplicate records based on geohash and year.
}
\examples{
\dontrun{
outqc_dup_check(dataset_a, dataset_b)
}

}
\author{
Pieter Provoost

Silas Principe
}
