% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqc_environmental.R
\name{outqc_env}
\alias{outqc_env}
\title{Detect geographical outliers based on distance between points}
\usage{
outqc_env(
  pts,
  variables = c("bathymetry", "sstemperature", "sssalinity", "shoredistance"),
  methods = c("iqr", "mad", "isoforest"),
  iqr_mltp = 3,
  mad_mltp = 6,
  isoforest_th = 0.8,
  iso_single = FALSE,
  limit_rem = 0.01
)
}
\arguments{
\item{pts}{the points for which detect outliers}

\item{variables}{the name of the variables that should be considered. One or more of
"bathymetry", "sstemperature", "sssalinity" or "shoredistance". Alternatively,
you can supply a SpatRaster with environmental layers. In that case, the function
will extract the information from those points.}

\item{methods}{which methods to use to detect outliers. One of "mdist_iqr",
"mdist_mad" or "isoforest" (see details)}

\item{iqr_mltp}{the multiplier of the interquantile range to use}

\item{mad_mltp}{the multiplier of the MAD to use}

\item{isoforest_th}{the threshold for the isoforest detection}

\item{iso_single}{if \code{TRUE}, then the Isolation Forest algorithm will run in
single variable mode, across each variable. If \code{FALSE} (default), it will
consider all variables together in a single model.}

\item{limit_rem}{a number between 0 and 1 limiting the percentage of points that can
be considered outliers. The default value (0.01) says that no more than 1\% of points
can be outliers}
}
\value{
a \code{matrix} with columns equal the number of methods * variables used. Each line is
one observation; values of 0 indicates normal values, while 1 indicates outliers.
}
\description{
Detect geographical outliers based on distance between points
}
\details{
Methods available are:
\itemize{
\item iqr: see if the observation is within \code{iqr_mltp} IQRs from the first and third quartile
\item mad: see if the observation is within \code{mad_mltp} MADs from the distance to the median value
\item isoforest: uses an extended isolation forest model (or single model, if \code{iso_single = TRUE})
to detect outliers. The higher the value of the output of such model, the more
likely it is to be an outlier. What value to use as a
threshold is difficulty to compute, as we don't have any control information
of what is indeed an outlier (otherwise we could have removed it!). After
tests with simulated and real data, the 0.8 value seems to be a reasonable
threshold.
}

If variables names are supplied, the \code{\link[obistools:lookup_xy]{obistools::lookup_xy()}} function is used
to get the environmental information.

\code{NA} values are not considered in the calculations, and are marked as NA
in the final outputs.
}
\examples{
\dontrun{

}
}
