% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqc_geographic.R
\name{outqc_geo}
\alias{outqc_geo}
\title{Detect geographical outliers based on distance between points}
\usage{
outqc_geo(
  pts,
  dist_folder = "distances",
  methods = c("mdist_iqr", "mdist_mad", "isoforest"),
  k = 15,
  iqr_mltp = 3,
  mad_mltp = 6,
  isoforest_th = 0.8,
  mdist_method = "mean",
  limit_rem = 0.01,
  parallel = TRUE,
  mc_cores = NULL
)
}
\arguments{
\item{pts}{the points for which detect outliers}

\item{dist_folder}{the folder where the distance layers produced with \code{\link[=outqc_get_distances]{outqc_get_distances()}}
are stored}

\item{methods}{which methods to use to detect outliers. One of "mdist_iqr",
"mdist_mad" or "isoforest" (see details)}

\item{k}{the number of nearest points to consider when computing the metrics.
If \code{NULL}, then nrow(pts) - 1 is used.}

\item{iqr_mltp}{the multiplier of the interquantile range to use}

\item{mad_mltp}{the multiplier of the MAD to use}

\item{isoforest_th}{the threshold for the isoforest detection}

\item{mdist_method}{which method to use to get the mean distance.
Usually mean, but can be "median"}

\item{limit_rem}{a number between 0 and 1 limiting the percentage of points that can
be considered outliers. The default value (0.01) says that no more than 1\% of points
can be outliers}

\item{parallel}{enable parallel computation on the extraction of values from
the distance layers (see \code{\link[=outqc_query_distances]{outqc_query_distances()}})}

\item{mc_cores}{number of cores to be used by parallel extraction}
}
\value{
a \code{matrix} with columns equal the number of methods used. Each line is
one observation; values of 0 indicates normal values, while 1 indicates outliers.
}
\description{
Detect geographical outliers based on distance between points
}
\details{
Methods available are:
\itemize{
\item mdist_iqr: see if the observation is within \code{iqr_mltp} IQRs from the first and third quartile
\item mdist_mad: see if the observation is within \code{mad_mltp} MADs from the distance to the median value
\item isoforest: uses an extended isolation forest model to detect outliers. The higher the value of the output
of such model, the more likely it is to be an outlier. What value to use as a threshold is difficulty
to compute, as we don't have any control information of what is indeed an outlier (otherwise we could have removed it!).
After tests with simulated and real data, the 0.8 value seems to be a reasonable threshold.
}
}
\examples{
\dontrun{

}
}
