% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqc_get_info.R
\name{outqc_get_base}
\alias{outqc_get_base}
\title{Get base raster for spatial outlier calculation}
\usage{
outqc_get_base(resolution = 1, maskpol = NULL)
}
\arguments{
\item{resolution}{a numeric value indicating the resolution of the base raster (in degrees).}

\item{maskpol}{a shapefile (sf or terra::vect) that will be used as the mask for the barriers.
Usually a world continents shapefile. If \code{NULL}, a world shapefile will be downloaded from
NaturalEarth.}
}
\value{
a SpatRaster, with values 1 in all sea areas and NA on land.
}
\description{
Get base raster for spatial outlier calculation
}
\examples{
\dontrun{
base <- outqc_get_base()
}
}
