% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqc_get_info.R
\name{outqc_get_distances}
\alias{outqc_get_distances}
\title{Get distances between all cells in a base raster, considering the continents
as barriers.}
\usage{
outqc_get_distances(
  base,
  target = NULL,
  agg_res = NULL,
  try_closest = TRUE,
  outfolder = "distances",
  skip_exists = TRUE,
  mc_cores = NULL,
  int_comp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{base}{a base raster to use for calculating distances. Can be produced
with \code{\link[=outqc_get_base]{outqc_get_base()}}}

\item{target}{optional, a dataframe containing the target x/y coordinates. If
NULL, the target will be all cells.}

\item{agg_res}{an optional resolution to first aggregate the points. This can
considerably speed up computation. Should be expressed in degrees.
This same value will be used to save a coarser resolution file. If \code{target}
is not provided and \code{agg_res} is, then the base file is aggregated and
all the cells for the coarser resolution are used (see details).}

\item{try_closest}{if \code{TRUE}, if the point lies in an invalid (empty) cell,
the function will try to find the closest valid point. It will look into the 8
adjacent cells. If there is still not one valid, this point is ignored.
Only relevant if target is supplied.}

\item{outfolder}{the folder where the files will be saved. If not existent,
it will be created.}

\item{skip_exists}{if TRUE, it skip cells that were already calculated. Just
set this to FALSE if you need to recalculate it.}

\item{mc_cores}{number of cores to use for parallel computation. Ignored on
non-UNIX systems.}

\item{int_comp}{if \code{TRUE} (default), then values are first converted to
integers and then the raster is saved with datatype "INT2U", what considerably
reduce file size. See \code{\link[terra:writeRaster]{terra::writeRaster()}} for more details.}

\item{verbose}{enable messages}
}
\value{
Files are saved on the disk.
}
\description{
This function will return the distances between each cell n and
all the other N(raster) cells. The distance is computed considering
continents as barriers.
\subsection{Choosing the resolution}{

Choosing a very coarse resolution will be problematic, as this will cause some
barriers to be artificially created (e.g. small straits may become a barrier,
even if there is a passage). A recommended resolution is of 0.1 degree for the
base raster. At the same time, this will cause the computation to be much slower
while also increasing the size of saved cells. In this case, it's interesting to
aggregate the points to a coarser grid (e.g. 0.5 degrees). This will considerably
improve computational time while also reducing the size of save files.
}

\subsection{Intended use}{

This function will save files with the distance for cell i to all other cells on
the raster (according to the desired resolution). And why to save? Because later you
can query the distances. The flux is for a set of points z, get all distances from
z{i} to z{n-i}. Then query the distances to perform outlier calculations. When you run
this function on the next dataset, it will run faster because potentially some of the cells
were already accounted for in the previous calculation. Thus, there is a time improvement each
time the function is run, up to the point that all cells are calculated.

Alternatively you can pre-calculate the distance to all cells by not supplying the
\code{target}.
}
}
\examples{
\dontrun{
# Get distance between cells in a 0.25 degrees resolution
outqc_get_distances(outqc_get_base(resolution = 0.25))
}

}
