% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqc_sac.R
\name{outqc_sac_mantel}
\alias{outqc_sac_mantel}
\title{Calculates spatial autocorrelation between points using ecospat method}
\usage{
outqc_sac_mantel(
  pts,
  env_layers,
  autocor_classdist = 5,
  autocor_maxdist = 200,
  autocor_signif = 0.05,
  return_pruned = TRUE,
  prune_threshold = 10,
  plot_result = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pts}{a \code{data.frame} with the coordinates to the points. Should
have at least two columns indicating the longitude and latitude (in that
order)}

\item{env_layers}{environmental layers in SpatRaster format}

\item{autocor_classdist}{the steps (in km) to calculate the autocorrelation}

\item{autocor_maxdist}{the maximum distance (in km) to which calculate the
autocorrelation}

\item{autocor_signif}{the significance level}

\item{return_pruned}{if \code{TRUE}, it will return the pruned records (if
the first non-correlated distance is higher or equal than \code{prune_threshold},
otherwise the original dataset is returned)}

\item{prune_threshold}{the minimum distance (in km) to use for prunning the
records}

\item{plot_result}{plot a graphical representation of the results}

\item{verbose}{print control messages}
}
\value{
the first non-correlated distance (in km) or prunned records
}
\description{
This function calculates the spatial autocorrelation between a set of points and,
optionally, prune the records based on the minimum non-significant distance.
This function differs from the \code{\link[=outqc_sac]{outqc_sac()}} in that it uses the \code{\link[ecospat:ecospat.mantel.correlogram]{ecospat::ecospat.mantel.correlogram()}}
method.
}
\details{
For each distance class, mantel r value is calculated. This procedure finds the minimum
non-significant autocorrelated distance which is optionally used to prune the
occurrence records

Note that if the minimum distance between points is higher than the maximum
distance being considered (\code{autocor_maxdist}), the function will abort
as it's impossible to assess the SAC. In that case, it makes no sense in prunning
the points.

If \code{return_pruned=TRUE}, the function \code{\link[spThin:thin]{spThin::thin()}} is used.
}
\examples{
\dontrun{
new_pts <- outqc_sac_mantel(pts, env)
}
}
\references{
Legendre, P. and M.J. Fortin. 1989. Spatial pattern and ecological analysis. Vegetatio, 80, 107-138.

\url{https://jorgemfa.medium.com/reducing-spatial-autocorrelation-in-species-distribution-models-fe84d4269cee}

Boavida, J., Assis, J., Silva, I. et al. (2016) Overlooked
habitat of a vulnerable gorgonian revealed in the Mediterranean and Eastern
Atlantic by ecological niche modelling. Scientific Reports. 6, 36460.
}
\seealso{
\code{\link[ecospat:ecospat.mantel.correlogram]{ecospat::ecospat.mantel.correlogram()}}
}
\author{
Original version by Jorge Assis (jmassis@ualg.pt), adapted to use terra functions.
}
