% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_tools.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Plot a Leaflet interactive map}
\usage{
plot_leaflet(layer, pts = NULL, map_palette = "Blues", points_color = "yellow")
}
\arguments{
\item{layer}{a SpatRaster with the prediction or environmental layer.
This can have one or more layers, preferentially named (otherwise, the
function will assign names as 'Layer 1', 'Layer 2', and so on)}

\item{pts}{an optional data.frame containing the geographic coordinates for points (two columns, x and y order)}

\item{map_palette}{the color palette to be used for plotting. See \code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}} for more details}

\item{points_color}{the color to be used fot the points, if supplied}
}
\value{
nothing, ploted map
}
\description{
Plot a Leaflet interactive map
}
\examples{
#' \dontrun{
plot_leaflet(sdm_prediction)
}
}
