% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recent_file}
\alias{recent_file}
\title{Retrieve the most recent file}
\usage{
recent_file(path, name, inverse = FALSE, by_date = FALSE)
}
\arguments{
\item{path}{path to search for the file}

\item{name}{name of the file to be searched (uses \code{\link[=grepl]{grepl()}})}

\item{inverse}{if \code{TRUE} return the oldest one}

\item{by_date}{if \code{TRUE} uses the last modified date instead}
}
\value{
file path
}
\description{
Retrieve the most recent file
}
\details{
In this project, many files are named with the date on the YMD format, like:
'all_splist_20240723.csv'. If \code{by_date} is \code{FALSE}, the function will simply
order the files (because is alphabetic, the most recent date will come first)
If the target file is not in that format, then it is better to use \code{by_date}
}
\examples{
\dontrun{
recent_file("data", "all_splist")
}
}
