% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{save_sdm}
\alias{save_sdm}
\title{Save the SDM model/metrics returned by the SDM modules}
\usage{
save_sdm(model, what, where)
}
\arguments{
\item{model}{an sdm_result object returned by one of the SDM modules}

\item{what}{what to save (e.g. model, cv_metrics)}

\item{where}{where to save (path)}
}
\value{
nothing, just saved objects
}
\description{
Save the SDM model/metrics returned by the SDM modules
}
\examples{
\dontrun{
save_sdm(my_model, "model", "models_folder/folder")
}
}
