% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_fit}
\alias{sdm_fit}
\title{Fit (tune) species distribution model}
\usage{
sdm_fit(
  sdm_data,
  sdm_method = "maxent",
  options = NULL,
  tune_blocks = "spatial_grid",
  metric = "cbi",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sdm_data}{\code{sdm_dat} object containing occurrences and environmental data}

\item{sdm_method}{algorithm for model fiting. See \code{\link[=sdm_options]{sdm_options()}} for available}

\item{options}{list with options from \code{\link[=sdm_options]{sdm_options()}} to override the default}

\item{tune_blocks}{which blocks to use for tuning}

\item{metric}{which metric to use for model optimization}

\item{verbose}{if \code{TRUE} print messages}

\item{...}{additional parameters (none implemented)}
}
\value{
\code{sdm_result} object containing the model
}
\description{
Fit (tune) species distribution model
}
\examples{
\dontrun{
sdm_species <- sdm_fit(sp_data)
}
}
