% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_module_esm}
\alias{sdm_module_esm}
\title{Fit Ensemble of Small Models using maxnet}
\usage{
sdm_module_esm(
  sdm_data,
  options = NULL,
  verbose = TRUE,
  tune_blocks = "spatial_grid",
  metric = "cbi"
)
}
\arguments{
\item{sdm_data}{\code{sdm_dat} object containing occurrences and environmental data}

\item{options}{list with options from \code{\link[=sdm_options]{sdm_options()}} to override the default}

\item{verbose}{if \code{TRUE} display messages}

\item{tune_blocks}{which blocks to use for cross-validation}

\item{metric}{which metric to optimize in tuning using cross-validation}
}
\value{
fitted model (\code{sdm_result} object)
}
\description{
Fit Ensemble of Small Models using maxnet
}
\examples{
\dontrun{
sdm_species <- sdm_module_esm(sp_data)
}
}
\references{
Breiner, F. T., Nobis, M. P., Bergamini, A., & Guisan, A. (2018).
Optimizing ensembles of small models for predicting the distribution of species
with few occurrences. In N. Isaac (Ed.), Methods in Ecology and Evolution
(Vol. 9, Issue 4, pp. 802–808). Wiley. https://doi.org/10.1111/2041-210x.12957
}
