% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_module_lgbm}
\alias{sdm_module_lgbm}
\title{Fit LightGBM (gradient-boosting machine) SDM}
\usage{
sdm_module_lgbm(
  sdm_data,
  options = NULL,
  verbose = TRUE,
  tune_blocks = "spatial_grid",
  metric = "cbi"
)
}
\arguments{
\item{sdm_data}{\code{sdm_dat} object containing occurrences and environmental data}

\item{options}{list with options from \code{\link[=sdm_options]{sdm_options()}} to override the default}

\item{verbose}{if \code{TRUE} display messages}

\item{tune_blocks}{which blocks to use for cross-validation}

\item{metric}{which metric to optimize in tuning using cross-validation}
}
\value{
fitted model (\code{sdm_result} object)
}
\description{
Fit LightGBM (gradient-boosting machine) SDM
}
\examples{
\dontrun{
sdm_species <- sdm_module_lgbm(sp_data)
}
}
