% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_module_maxent}
\alias{sdm_module_maxent}
\title{SDM module: MAXENT (Maximum Entropy)}
\usage{
sdm_module_maxent(
  sdm_data,
  method = "maxnet",
  tune_blocks = "spatial_grid",
  blocks_all = FALSE,
  maxent_pred_layers = NULL
)
}
\arguments{
\item{sdm_data}{object of class sdm_dat returned by \code{\link[=mp_prepare_data]{mp_prepare_data()}}}

\item{method}{either "maxnet" to use the \code{\link[maxnet:maxnet]{maxnet::maxnet()}} implementation or
"maxent" to use the Java implementation (see note).}

\item{tune_blocks}{the name of the blocks (added with \code{\link[=mp_prepare_blocks]{mp_prepare_blocks()}})
which should be used for tuning the models}

\item{blocks_all}{if \code{TRUE}, then after tuning the model, it will
produce cross-validated metrics for all types of blocks available (e.g.
latitudinal)}

\item{maxent_pred_layers}{optional. If method is "maxent", there is no predict
method available. In this case, if you want the predictions, you need to supply
the prediction layers (in SpatRaster format) here}
}
\value{
an sdm_result object (a list) containing:
\itemize{
\item name = name of method
\item model = the model object
\item timings = a character vector of the timings for each step
\item parameters = the best tuned parameters
\item cv_metrics = the cross-validated metrics
\item full_metrics = the metrics of the final model fitted using the whole data.
Just for reference, as this is not a fair scenario
(same data used for training and testing).
\item eval_metrics = metrics for the evaluation dataset (if this is supplied).
}
}
\description{
SDM module: MAXENT (Maximum Entropy)
}
\note{
The dismo implementation of the Maxent (through rJava) is causing R to crash.
Thus we use here a workaround which call the java program through a system call.
It's advisable to use the maxnet implementation which produces basically identical
outputs and have a predict method.
}
\examples{
\dontrun{
sdm_module_maxent(species_data)
}
}
