% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_module_rf}
\alias{sdm_module_rf}
\title{Fit Random Forest SDM using randomForest}
\usage{
sdm_module_rf(
  sdm_data,
  options = NULL,
  verbose = TRUE,
  tune_blocks = "spatial_grid",
  metric = "cbi"
)
}
\arguments{
\item{sdm_data}{\code{sdm_dat} object containing occurrences and environmental data}

\item{options}{list with options from \code{\link[=sdm_options]{sdm_options()}} to override the default}

\item{verbose}{if \code{TRUE} display messages}

\item{tune_blocks}{which blocks to use for cross-validation}

\item{metric}{which metric to optimize in tuning using cross-validation}
}
\value{
fitted model (\code{sdm_result} object)
}
\description{
Fit Random Forest SDM using randomForest
}
\details{
when method is "down-sampled", the internal down-sampling
(using \link[randomForest:randomForest]{randomForest::randomForest} argument \code{sampsize}) only opperates
on the absences. Thus, in the rare cases where there are more presences than
absences it will, instead, directly sample the presences to reduce its number
in order to achieve the same number of absences. This is only done at the
block level. If, in total, the number of presences is higher than absences,
then the function will fail and you should check the data.
}
\examples{
\dontrun{
sdm_species <- sdm_module_rf(sp_data)
}
}
