% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_multhypo}
\alias{sdm_multhypo}
\title{Test multiple hypothesis}
\usage{
sdm_multhypo(
  sdm_data,
  variables,
  fast_mode = TRUE,
  sdm_method = "maxent",
  compare_metric = "cbi",
  return_all = FALSE,
  verbose = TRUE,
  options = NULL,
  ...
)
}
\arguments{
\item{sdm_data}{\code{sdm_dat} object containing occurrences and environmental data}

\item{variables}{list of hypothesis}

\item{fast_mode}{if \code{TRUE}, only the first model is tuned, and subsequent hypothesis are
tested using the best tune parameters}

\item{sdm_method}{algorithm to fit}

\item{compare_metric}{which metric to compare between hypothesis}

\item{return_all}{if \code{TRUE} return all models (default is \code{FALSE})}

\item{verbose}{turn on messages}

\item{options}{options for model fitting}

\item{...}{additional parameters (none implemented)}
}
\value{
fitted model, with best model information
}
\description{
Test multiple hypothesis
}
\examples{
\dontrun{
best_hypothesis <- sdm_multhypo(sp_data, hypo_list)
}
}
