% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_options}
\alias{sdm_options}
\title{Retrieve options used for model tuning}
\usage{
sdm_options(sdm_method = NULL)
}
\arguments{
\item{sdm_method}{\code{NULL} for returning full list, or an algorithm name}
}
\value{
list of options for the chosen algorithm or all
}
\description{
Retrieve options used for model tuning
}
\examples{
# Print all information
sdm_options()

# Retrieve for a single algorithm
sdm_options("maxent")

# Change details of an algorithm
max_opt <- sdm_options("maxent")
max_opt$features <- c("lq", "lqh")

# Supply to fit function
\dontrun{
sdm_species <- sdm_fit(sp_data, options = max_opt)
}

}
