% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules_gee.R
\name{start_ee}
\alias{start_ee}
\title{Start rgee (Google Earth Engine) session}
\usage{
start_ee(...)
}
\arguments{
\item{...}{optional parameters passed to [rgee::ee_Authenticate())]

[rgee::ee_Authenticate())]: R:rgee::ee_Authenticate())}

\item{init_only}{if \code{TRUE}, just initialize session. Useful if you already authenticated but restarted the R session}
}
\value{
nothing, just starts session
}
\description{
Start rgee (Google Earth Engine) session
}
\details{
Soon more details.
}
\examples{
\dontrun{
start_ee()
}
}
