% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{variable_importance}
\alias{variable_importance}
\title{Assess the importance of variables for the model}
\usage{
variable_importance(model, sdm_data, iterations = 10)
}
\arguments{
\item{model}{}

\item{sdm_data}{}

\item{iterations}{}
}
\value{
a data.frame with variables importance
}
\description{
Assess the importance of variables for the model
}
\details{
This function calculates the importance of variables using the same principle
of \code{\link[randomForest:randomForest]{randomForest::randomForest()}} and \code{\link[biomod2:bm_VariablesImportance]{biomod2::bm_VariablesImportance()}}.

Once each time, a variable is shuffled and a new prediction is made. The importance of the variable
is given as 1-(correlation between original and new prediction). Each variable
is shuffled several times (controled by the argument \code{iterations}) and the
function returns the mean and standard deviation. The higher the score, higher is
the importance of the variable for the model.
}
\examples{
\dontrun{
variable_importance(model, sp_data)
}
}
