% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{variable_importance_esm}
\alias{variable_importance_esm}
\title{Assess the importance of variables for ESM model}
\usage{
variable_importance_esm(model, sdm_data, iterations = 10)
}
\arguments{
\item{model}{ESM models list}

\item{sdm_data}{sdm_data object}

\item{iterations}{number of iteration}
}
\value{
a data.frame with variables importance
}
\description{
Assess the importance of variables for ESM model
}
\details{
This function calculates the importance of variables using the same principle
of \code{\link[randomForest:randomForest]{randomForest::randomForest()}} and \code{\link[biomod2:bm_VariablesImportance]{biomod2::bm_VariablesImportance()}}.

Once each time, a variable is shuffled and a new prediction is made. The importance of the variable
is given as 1-(correlation between original and new prediction). Each variable
is shuffled several times (controled by the argument \code{iterations}) and the
function returns the mean and standard deviation. The higher the score, higher is
the importance of the variable for the model.
}
\examples{
\dontrun{
variable_importance_esm(model, sp_data)
}
}
