% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_tools.R
\name{gen_log}
\alias{gen_log}
\title{Generate log of modeling}
\usage{
gen_log(algos)
}
\arguments{
\item{algos}{algorithms that will be used to fit}
}
\value{
a list
}
\description{
Generate log of modeling
}
\details{
The following parameters are available in the list object:
\itemize{
\item taxonID
\item scientificName
\item group
\item model_date
\item model_acro
\item n_init_points = input number of initial points before autocorrelation
\item model_fit_points
\item model_eval_points
\item algorithms
\item algorithms_parameters = list with parameters used for model fitting (auto filled)
\item model_result = result of model fitting
\item model_bestparams = the best fitting parameters
\item model_details = list with the following parameters:
\itemize{
\item ecoregions
\item ecoregions_included
\item limited_by_depth = TRUE or FALSE, if it was limited by depth
\item depth_buffer
\item block_size
\item background_size
\item control_bias
\item hypothesis_tested
\item best_hypothesis
\item variables
}
\item model_good = which models performed well according to the metric and threshold
\item model_good_metric = metric used for defining "good" models
\item model_good_threshold = threshold used for defining "good" models
\item model_posteval = list with post-evaluation results
\item timings
\item obissdm_version = auto filled, version of the \code{obissdm} package
}

Other parameters can be added as needed.
}
\examples{
sdm_log <- gen_log(c("maxent", "brt"))
\dontrun{
save_log(sdm_log, "species_log.json")
}
}
\seealso{
\code{\link[=save_log]{save_log()}}
}
