% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_get.R
\name{mp_get_obis}
\alias{mp_get_obis}
\title{Retrieve occurrence records from OBIS and save in the standard format}
\usage{
mp_get_obis(
  sci_names,
  start_year = 1950,
  save_format = "parquet",
  mode = "single",
  save_acro = NULL
)
}
\arguments{
\item{sci_names}{a single value or a vector containing scientific names or, ideally, AphiaID (\code{numeric})}

\item{start_year}{the initial year from which occurrences should be get}

\item{save_format}{the format to save the files. Either "parquet" (recommended) or "csv"}

\item{mode}{mode of export. Either "single" (for single species) or "full" for multiple species. When using "full",
a single file containing all the records will be saved in the "data/raw" folder.}

\item{save_acro}{in case of full export, an acronym to be added to the save file. If NULL, "full" is used.}
}
\value{
Files saved
}
\description{
Retrieve occurrence records from OBIS and save in the standard format
}
\examples{
\dontrun{

}
}
