% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obissdm-package.R
\docType{package}
\name{obissdm-package}
\alias{obissdm}
\alias{obissdm-package}
\title{obissdm: Generate SDMs for MPA Europe project}
\description{
This package was created to hold the functions used to generate the SDMs for the MPA Europe project.
}
\author{
\strong{Maintainer}: Silas Principe \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
