% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_prepare_blocks.R
\name{plot_folds}
\alias{plot_folds}
\title{Plot CV folds}
\usage{
plot_folds(
  sdm_data,
  block_type = NULL,
  facet_run = FALSE,
  presence_only = FALSE,
  include_map = FALSE
)
}
\arguments{
\item{sdm_data}{an object of type \code{sdm_dat} produced with
\code{\link[=mp_prepare_data]{mp_prepare_data()}} and with cross-validation blocks assigned using
\code{\link[=mp_prepare_blocks]{mp_prepare_blocks()}}}

\item{block_type}{which block type should be ploted (one single name). If
NULL the first one will be ploted}

\item{facet_run}{if \code{TRUE}, then facets equal the number of folds will
be ploted, each showing which records will be used for training and which
for testing}

\item{presence_only}{if \code{TRUE}, then only the presence points are ploted}

\item{include_map}{if \code{TRUE}, a base map of the continents is added for
reference. The package \link{rnaturalearth} is needed for this}
}
\value{
nothing
}
\description{
Plot the cross-validation folds produced with \code{\link[=mp_prepare_blocks]{mp_prepare_blocks()}}.
}
\examples{
\dontrun{
plot_folds(sdm_data)
}
}
