% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_addons.R
\name{resp_curves}
\alias{resp_curves}
\title{Generate response curves for the SDM models}
\usage{
resp_curves(model, layers, sdm_data = NULL, vars = NULL)
}
\arguments{
\item{model}{an object of class \code{sdm_result} returned by one of the SDM
modules}

\item{layers}{environmental layers used to fit the model (SpatRaster format)}

\item{sdm_data}{optional, an \code{sdm_dat} object containing the fitting data.
If supplied, a column called \code{in_range} will be added, with \code{0} for values
out of the training range and \code{1} for those within the range.}

\item{vars}{optional, a vector with names of the variables for which the
response curves should be obtained.}
}
\value{
a data.frame containing predictions for the response curves
}
\description{
Generate response curves for the SDM models
}
\details{
A plot method is available for the object generated by this function
}
\examples{
\dontrun{

}
}
