% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_tools.R
\name{save_log}
\alias{save_log}
\alias{view_log}
\title{Save log object}
\usage{
save_log(log_object, file_path)

view_log(log_object)
}
\arguments{
\item{log_object}{log object generated with \code{\link[=gen_log]{gen_log()}}. For \code{\link[=view_log]{view_log()}}
can also be a path to a json file}

\item{file_path}{the path to save the file in .json format}
}
\value{
saved file
}
\description{
Save log object
}
\details{
You can quickly visualize the log files using \code{\link[=view_log]{view_log()}}
}
\examples{
\dontrun{
save_log(sdm_log, "species_log.json")
}
}
