% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_modules.R
\name{sdm_module_esm}
\alias{sdm_module_esm}
\title{Fit Ensemble of Small Models using maxnet}
\usage{
sdm_module_esm(
  sdm_data,
  options = NULL,
  verbose = TRUE,
  tune_blocks = "spatial_grid",
  metric = "cbi",
  esm_metric = "dsomers",
  esm_threshold = 0,
  remove_below = TRUE
)
}
\arguments{
\item{sdm_data}{\code{sdm_dat} object containing occurrences and environmental data}

\item{options}{list with options from \code{\link[=sdm_options]{sdm_options()}} to override the default}

\item{verbose}{if \code{TRUE} display messages}

\item{tune_blocks}{which blocks to use for cross-validation}

\item{metric}{which metric to optimize in tuning using cross-validation}

\item{esm_metric}{metric to filter bivariate models. One of "dsomer", which
follows original article using Somers'D, or "cbi".}

\item{esm_threshold}{value for filtering out models. Default is 0, which for
Somers'D is equivalent to remove models with AUC < 0.5}

\item{remove_below}{if \code{TRUE} remove models below the \code{esm_threshold}}
}
\value{
list of fitted bivariate models (\code{sdm_esm_result} object, with
\code{sdm_result} objects). The main object contains an attribute \code{scores} which
provide easy access to the scores for model weighting.
}
\description{
Fit Ensemble of Small Models using maxnet
}
\note{
Avoid using more than 50000 background points for ESM.
}
\examples{
\dontrun{
sdm_species <- sdm_module_esm(sp_data)
}
}
\references{
Breiner, F. T., Nobis, M. P., Bergamini, A., & Guisan, A. (2018).
Optimizing ensembles of small models for predicting the distribution of species
with few occurrences. In N. Isaac (Ed.), Methods in Ecology and Evolution
(Vol. 9, Issue 4, pp. 802–808). Wiley. https://doi.org/10.1111/2041-210x.12957
}
